taskweek3-5/requirements.txt
pymupdf
qdrant-client
fastapi
uvicorn
pika
psycopg2-binary
sentence-transformers
python-multipart
python-dotenv
